Move more of the logic to the runtime
